requirements.txt
llama3
evolutionapi
python-dotenv
